SSHOTPLAT-4292: sbl(3.21.13): Modify pre-up check threshold for fabric links (#427)
Changes to use aggressive thresholds when saved tuning parameters are used
for optimal performance. Saved tuning parameters would be discarded and the
link will be retuned if the BER is not up to the standards.

It should be noted that this change only modifies the pre-up check thresholds
when saved tuning parameters are used.